requirements.txt
requests
pystyle
pyyaml